LayoutTests/platform/mac/http/tests/media/media-source/SourceBuffer-abort-readyState-expected.txt


FAIL SourceBuffer#abort() (video/webm; codecs="vorbis,vp8") : If the readyState attribute of the parent media source is not in the "open" state then throw an INVALID_STATE_ERR exception and abort these steps. assert_unreached: Browser doesn't support the MIME used in this test: video/webm; codecs="vorbis,vp8" Reached unreachable code(stack: assert@http://127.0.0.1:8000/w3c/resources/testharness.js:2060:37
assert_unreached@http://127.0.0.1:8000/w3c/resources/testharness.js:1056:16
http://127.0.0.1:8000/media/media-source/SourceBuffer-abort-readyState.html:20:29
step@http://127.0.0.1:8000/w3c/resources/testharness.js:1160:30
isTypeSupported@http://127.0.0.1:8000/media/media-source/SourceBuffer-abort-readyState.html:19:18
isTypeSupported@[native code]
http://127.0.0.1:8000/media/media-source/SourceBuffer-abort-readyState.html:42:36
step@http://127.0.0.1:8000/w3c/resources/testharness.js:1160:30
async_test@http://127.0.0.1:8000/w3c/resources/testharness.js:460:26
mediaTest@http://127.0.0.1:8000/media/media-source/SourceBuffer-abort-readyState.html:41:15
global code@http://127.0.0.1:8000/media/media-source/SourceBuffer-abort-readyState.html:67:14)
PASS SourceBuffer#abort() (video/mp4) : If the readyState attribute of the parent media source is not in the "open" state then throw an INVALID_STATE_ERR exception and abort these steps.
